requirements.txt
annotated-types==0.5.0
anyio==3.7.1
fastapi==0.103.1
httptools==0.6.0
idna==3.4
option==2.1.0
pydantic==2.4.0
pydantic_core==2.10.0
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
typing_extensions==4.8.0
uvicorn==0.23.2
websockets==11.0.3
numpy==1.24.4
nanoid==2.0.0
pandas==2.0.3